Added curvature regularization to cashocs
This is now validated and works as expected. However, non-smooth geometries might be an issue for this.
Additionally, fixed a small typo for the shape gradient tests, which caused the direction to be subtracted instead of added. Since the problem is, however, linear, this does not make an actual difference for the result.