Started code refactoring
Goal: Better readability, better maintainability.
This update drops the support for the scaling functionality, which is no longer supported.
Maybe the scaling can be reimplemented in a different form in the future, but as it was now, it was rather intransparent and could not work properly with the constraints module.